requirements.txt
pandas
matplotlib
plotly
streamlit
statsmodels==0.14.0
joblib==1.2.0
pillow==10.2.0
altair==5.0.1
altair-saver==0.5.0
st-pages==0.4.5
kaleido==0.2.1